+++ src/containerapp/azext_containerapp/tests/latest/test_containerapp_session_code_interpreter_files.py
@@ -10,14 +10,13 @@ from azure.cli.testsdk.scenario_tests import AllowLargeResponse
 from azure.cli.testsdk import (ScenarioTest, ResourceGroupPreparer, JMESPathCheck)
 
 from .common import (TEST_LOCATION)
-from .utils import create_containerapp_env
 
 TEST_DIR = os.path.abspath(os.path.join(os.path.abspath(__file__), '..'))
 
-class ContainerAppSessionCodeInterperterTests(ScenarioTest):
+class ContainerAppSessionCodeInterperterFilesTests(ScenarioTest):
     @AllowLargeResponse(8192)
     @ResourceGroupPreparer()
-    def test_containerapp_session_code_interpreter_e2e(self, resource_group):
+    def test_containerapp_session_code_interpreter_files_e2e(self, resource_group):
         location = TEST_LOCATION
         self.cmd('configure --defaults location={}'.format(location))
 
@@ -27,8 +26,8 @@ class ContainerAppSessionCodeInterperterTests(ScenarioTest):
 
         # Create JupyterPython SessionPool
         sessionpool_name_python = self.create_random_name(prefix='spjupyterpython', length=24)
-        self.cmd('containerapp sessionpool create -g {} -n {} --cooldown-period {} -l {}'.format(
-            resource_group, sessionpool_name_python, 300, location), checks=[
+        self.cmd('containerapp sessionpool create -g {} -n {} --cooldown-period {}'.format(
+            resource_group, sessionpool_name_python, 300), checks=[
             JMESPathCheck('name', sessionpool_name_python),
             JMESPathCheck('properties.containerType', "PythonLTS"),
             JMESPathCheck('properties.provisioningState', "Succeeded"),
@@ -53,53 +52,61 @@ class ContainerAppSessionCodeInterperterTests(ScenarioTest):
         ])
 
         # upload a file also add session pool location
-        txt_file = os.path.join(TEST_DIR, 'cert.txt')
-        self.cmd('containerapp session code-interpreter upload-file -n {} -g {} --identifier {} --filepath "{}" --session-pool-location {}'.format(
+        txt_file_name = 'cert.txt'
+        txt_file = os.path.join(TEST_DIR, txt_file_name)
+        path = '/'
+        self.cmd('containerapp session code-interpreter upload-file -n {} -g {} --identifier {} --filepath "{}" --path {} --session-pool-location {}'.format(
             sessionpool_name_python,
             resource_group,
             identifier_name,
             txt_file,
-            location),
+            path,
+            TEST_LOCATION),
             checks=[
-            JMESPathCheck('name', 'cert.txt'),
+            JMESPathCheck('name', txt_file_name),
         ])
 
         # list files
-        files_list = self.cmd("containerapp session code-interpreter list-files -n {} -g {} --identifier {}".format(
+        files_list = self.cmd("containerapp session code-interpreter list-files -n {} -g {} --identifier {} --path {}".format(
             sessionpool_name_python,
             resource_group,
-            identifier_name)).get_output_in_json()
+            identifier_name,
+            path)).get_output_in_json()
         self.assertTrue(len(files_list["value"]) == 1)
 
         # check content
-        file_content = self.cmd("containerapp session code-interpreter show-file-content -n {} -g {} --identifier {} --filename {}".format(
+        file_content = self.cmd("containerapp session code-interpreter show-file-content -n {} -g {} --identifier {} --filename {} --path {}".format(
             sessionpool_name_python,
             resource_group,
             identifier_name,
-            "cert.txt")).get_output_in_json()
+            txt_file_name,
+            path)).get_output_in_json()
         self.assertTrue(file_content == '\"testing\"')
 
         # check metadata
-        self.cmd("containerapp session code-interpreter show-file-metadata -n {} -g {} --identifier {} --filename {}".format(
+        self.cmd("containerapp session code-interpreter show-file-metadata -n {} -g {} --identifier {} --filename {} --path {}".format(
             sessionpool_name_python,
             resource_group,
             identifier_name,
-            "cert.txt"),
+            txt_file_name,
+            path),
             checks=[
-            JMESPathCheck('name', 'cert.txt'),
+            JMESPathCheck('name', txt_file_name),
         ])
 
         # delete file
-        self.cmd("containerapp session code-interpreter delete-file -n {} -g {} --identifier {} --filename {} --yes".format(
+        self.cmd("containerapp session code-interpreter delete-file -n {} -g {} --identifier {} --filename {} --path {} --yes".format(
             sessionpool_name_python,
             resource_group,
             identifier_name,
-            "cert.txt"
+            txt_file_name,
+            path
             ))
-        files_list = self.cmd("containerapp session code-interpreter list-files -n {} -g {} --identifier {}".format(
+        files_list = self.cmd("containerapp session code-interpreter list-files -n {} -g {} --identifier {} --path {}".format(
             sessionpool_name_python,
             resource_group,
-            identifier_name
+            identifier_name,
+            path
             )).get_output_in_json()
         self.assertTrue(len(files_list["value"]) == 0)